Remove temporary refactor subdirectory.
