+++ src/rinqamqp/internal/amqptest/connection.go
