+++ src/rinqamqp/internal/amqpx/headers.go
