+++ src/rinqamqp/internal/marshaling/trace.go
