+++ src/rinqamqp/internal/marshaling/trace_test.go
@@ -3,8 +3,8 @@ package marshaling_test
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/rinq/rinq-go/src/rinqamqp/internal/refactor/amqptest"
-	. "github.com/rinq/rinq-go/src/rinqamqp/internal/refactor/marshaling"
+	"github.com/rinq/rinq-go/src/rinqamqp/internal/amqptest"
+	. "github.com/rinq/rinq-go/src/rinqamqp/internal/marshaling"
 	"github.com/streadway/amqp"
 )
