+++ src/rinqamqp/internal/notifications/marshallog.go
@@ -4,7 +4,7 @@ import (
 	"github.com/jmalloc/twelf/src/twelf"
 	"github.com/rinq/rinq-go/src/internal/transport"
 	"github.com/rinq/rinq-go/src/rinq/ident"
-	"github.com/rinq/rinq-go/src/rinqamqp/internal/refactor/marshaling"
+	"github.com/rinq/rinq-go/src/rinqamqp/internal/marshaling"
 	"github.com/streadway/amqp"
 )
